pyproject.toml
[project]
name = "soil-moisture-monitor"
version = "0.1.0"
description = ""
authors = [
    {name = "DalleAntoine-99",email = "antoine.dalle59@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = ["torch (>=2.10.0,<3.0.0)", "transformers (>=5.0.0,<6.0.0)", "numpy (>=2.4.1,<3.0.0)", "scikit-learn (>=1.8.0,<2.0.0)", "pillow (>=12.1.0,<13.0.0)", "tqdm (>=4.67.1,<5.0.0)", "huggingface-hub (>=1.3.4,<2.0.0)", "torchvision (>=0.25.0,<0.26.0)"]


[tool.poetry.dependencies]
python = ">=3.11,<3.15"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "huggingface-hub (>=1.3.4,<2.0.0)"
]
